int -> float

+++ average_calculator.py
@@ -1,7 +1,7 @@
 print("/!\ Séparez les notes par des espaces et remplacez les virgules par des points /!\ ")
 notes  = input("Entrez vos notes : ").split()
 
-notes  = [int(i) for i in notes]
+notes  = [float(i) for i in notes]
 
 result = sum(notes) / len(notes)
 print(result)
